+++ plugin-dev/Source/Sentry/Private/HAL/PlatformSentryBreadcrumb.h
@@ -6,6 +6,8 @@
 #include "Apple/SentryBreadcrumbApple.h"
 #elif USE_SENTRY_NATIVE
 #include "GenericPlatform/GenericPlatformSentryBreadcrumb.h"
+#else
+#include "Interface/SentryBreadcrumbInterface.h"
 #endif
 
 static TSharedPtr<ISentryBreadcrumb> CreateSharedSentryBreadcrumb()