+++ plugin-dev/Source/Sentry/Private/HAL/PlatformSentryId.h
@@ -4,17 +4,21 @@
 #include "Android/SentryIdAndroid.h"
 #elif PLATFORM_APPLE
 #include "Apple/SentryIdApple.h"
-#else
+#elif USE_SENTRY_NATIVE
 #include "GenericPlatform/GenericPlatformSentryId.h"
+#else
+#include "Interface/SentryIdInterface.h"
 #endif
 
 static TSharedPtr<ISentryId> CreateSharedSentryId()
 {
 #if PLATFORM_ANDROID
-    return MakeShareable(new SentryIdAndroid);
+	return MakeShareable(new SentryIdAndroid);
 #elif PLATFORM_APPLE
-    return MakeShareable(new SentryIdApple);
+	return MakeShareable(new SentryIdApple);
+#elif USE_SENTRY_NATIVE
+	return MakeShareable(new FGenericPlatformSentryId);
 #else
-    return MakeShareable(new FGenericPlatformSentryId);
+	return nullptr;
 #endif
 }