+++ plugin-dev/Source/Sentry/Private/HAL/PlatformSentryTransactionContext.h
@@ -4,8 +4,10 @@
 #include "Android/SentryTransactionContextAndroid.h"
 #elif PLATFORM_APPLE
 #include "Apple/SentryTransactionContextApple.h"
-#else
+#elif USE_SENTRY_NATIVE
 #include "GenericPlatform/GenericPlatformSentryTransactionContext.h"
+#else
+#include "Interface/SentryTransactionContextInterface.h"
 #endif
 
 static TSharedPtr<ISentryTransactionContext> CreateSharedSentryTransactionContext(const FString& Name, const FString& Operation)
@@ -14,7 +16,9 @@ static TSharedPtr<ISentryTransactionContext> CreateSharedSentryTransactionContex
 	return MakeShareable(new SentryTransactionContextAndroid(Name, Operation));
 #elif PLATFORM_APPLE
 	return MakeShareable(new SentryTransactionContextApple(Name, Operation));
-#else
+#elif USE_SENTRY_NATIVE
 	return MakeShareable(new FGenericPlatformSentryTransactionContext(Name, Operation));
+#else
+	return nullptr;
 #endif
 }